+++ types/resources/messages.d.ts
@@ -90,6 +90,7 @@ export class MessageInterface extends PlivoResourceInterface {
 	 * @param {string} [optionalParams.method] The method used to call the url. Defaults to POST.
 	 * @param {list} [optionalParams.media_urls] For sending mms, specify the media urls in list of string
 	 * @param {boolean} [optionalParams.log] If set to false, the content of this message will not be logged on the Plivo infrastructure and the dst value will be masked (e.g., 141XXXXX528). Default is set to true.
+	 * @param {Template} [optionalParams.template] For sending templated whatsapp messages.
 	 * @promise {object} return {@link PlivoGenericMessage} object if success
 	 * @fail {Error} return Error
 	 */
@@ -112,6 +113,7 @@ export class MessageInterface extends PlivoResourceInterface {
 	 * @param {string} [optionalParams.method] The method used to call the url. Defaults to POST.
 	 * @param {boolean} [optionalParams.log] If set to false, the content of this message will not be logged on the Plivo infrastructure and the dst value will be masked (e.g., 141XXXXX528). Default is set to true.
 	 * @param {Array} [optionalParams.media_urls] For sending mms, specify the media urls in list of string
+	 * @param {Template} [optionalParams.template] For sending templated whatsapp messages.
 	 * @promise {object} return {@link MessageResponse} object if success
 	 * @fail {Error} return Error
 	 */